python/requirements.txt
openai==1.12.0
anthropic==0.18.1
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.6.0
httpx==0.27.0
tiktoken==0.5.2
tenacity==8.2.3
aiohttp==3.9.3
sqlite3
json5==0.9.20